Export java edition mcdoc attributes separately
