+++ packages/java-edition/src/index.ts
@@ -16,6 +16,7 @@ import {
 	symbolRegistrar,
 } from './dependency/index.js'
 import * as jeJson from './json/index.js'
+import { registerMcdocAttributes, registerPackFormatAttribute } from './mcdocAttributes.js'
 import * as jeMcf from './mcfunction/index.js'
 
 export * as binder from './binder/index.js'
@@ -121,82 +122,8 @@ export const initialize: core.ProjectInitializer = async (ctx) => {
 				&& !n.symbol?.path[0]?.startsWith('::minecraft')),
 	})
 
-	mcdoc.runtime.registerAttribute(meta, 'since', mcdoc.runtime.attribute.validator.string, {
-		filterElement: (config, ctx) => {
-			if (!config.startsWith('1.')) {
-				ctx.logger.warn(`Invalid mcdoc attribute for "since": ${config}`)
-				return true
-			}
-			return ReleaseVersion.cmp(release, config as ReleaseVersion) >= 0
-		},
-	})
-	mcdoc.runtime.registerAttribute(meta, 'until', mcdoc.runtime.attribute.validator.string, {
-		filterElement: (config, ctx) => {
-			if (!config.startsWith('1.')) {
-				ctx.logger.warn(`Invalid mcdoc attribute for "until": ${config}`)
-				return true
-			}
-			return ReleaseVersion.cmp(release, config as ReleaseVersion) < 0
-		},
-	})
-	mcdoc.runtime.registerAttribute(
-		meta,
-		'deprecated',
-		mcdoc.runtime.attribute.validator.optional(mcdoc.runtime.attribute.validator.string),
-		{
-			mapField: (config, field, ctx) => {
-				if (config === undefined) {
-					return { ...field, deprecated: true }
-				}
-				if (!config.startsWith('1.')) {
-					ctx.logger.warn(`Invalid mcdoc attribute for "deprecated": ${config}`)
-					return field
-				}
-				if (ReleaseVersion.cmp(release, config as ReleaseVersion) >= 0) {
-					return { ...field, deprecated: true }
-				}
-				return field
-			},
-		},
-	)
-	const packFormats = new Map<number, McmetaVersion>()
-	for (const version of versions) {
-		if (version.type === 'release' && !packFormats.has(version.data_pack_version)) {
-			packFormats.set(version.data_pack_version, version)
-		}
-	}
-	mcdoc.runtime.registerAttribute(meta, 'pack_format', () => undefined, {
-		checker: (_, typeDef) => {
-			if (typeDef.kind !== 'literal' || typeof typeDef.value.value !== 'number') {
-				return undefined
-			}
-			const target = typeDef.value.value
-			return (node, ctx) => {
-				const targetVersion = packFormats.get(target)
-				if (!targetVersion) {
-					ctx.err.report(
-						localize('java-edition.pack-format.unsupported', target),
-						node,
-						core.ErrorSeverity.Warning,
-					)
-				} else if (targetVersion.id !== release) {
-					ctx.err.report(
-						localize('java-edition.pack-format.not-loaded', target, release),
-						node,
-						core.ErrorSeverity.Warning,
-					)
-				}
-			}
-		},
-		numericCompleter: (_, ctx) => {
-			return [...packFormats.values()].map((v, i) => ({
-				range: core.Range.create(ctx.offset),
-				label: `${v.data_pack_version}`,
-				labelSuffix: ` (${v.id})`,
-				sortText: `${i}`.padStart(4, '0'),
-			} satisfies core.CompletionItem))
-		},
-	})
+	registerMcdocAttributes(meta, release)
+	registerPackFormatAttribute(meta, release, versions)
 
 	json.initialize(ctx)
 	jeJson.initialize(ctx)
